+++ include/global.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <string>
+#include "json.hpp"
+
+namespace global {
+
+    extern std::string github_url;
+    extern std::string app_version;
+    extern nlohmann::json app_config;
+    extern nlohmann::json template_info_config;
+    extern nlohmann::json template_variables_config;
+    extern std::string cache_container_name;
+}